+++ NEWS.md
@@ -1,9 +1,9 @@
 # discretewq (development version)
 
 * Added `DOP` water quality data.
-* Updated baystudy, DJFMP, EDSM, NCRO, SKT, and twentymm datasets.
-* Added bottom conductivity to `EDSM`
-* Added turbidity to `DJFMP`, `EDSM`, `SKT`, and `twentymm`
+* Updated baystudy, DJFMP, EDSM, NCRO, SKT, SLS, and twentymm datasets.
+* Added bottom conductivity to `EDSM` and `SLS`
+* Added turbidity to `DJFMP`, `EDSM`, `SKT`, `SLS`, and `twentymm`
 * Improved backend of data processing to add automation and data checks
 
 # discretewq 2.4.0